requirements.txt
pandas
matplotlib
psycopg2-binary
